Minimum fixups to make fusell usable on python 3
Which basically means a little bit of attention to the unicode<->str conflict.

This is far from ideal, but for now the filenames on the callbacks and internal functions are not encoded/decoded to unicode: the application should always deal with `bytes` instead of `str`.
This was done to minimize the fuzz and keep backward-compatibility