[fusell] Support for character encoding
The bytes<->unicode mismatch between Python 2 and 3 made fusell impossible to use on Py3 until now.

Previously, on PR #89, I made the minimal modifications necessary to have it working on Py3, but it basically left the duty of character encoding/decoding to the application.

Instead, this PR fully implements encoding/decoding of strings on the fuse_xxx, reply_readdir and reply_readlink functions, so that the applications only needs to handle unicode strings, similar to how it's done on fuse.py.