call native fuse_exit when a critical exception arises
BaseExceptions that are not an instance of Exception and all Exceptions
thrown from init are deemed critical. The python method fuse_exit() is
added to call the native fuse_exit method. The exception causing the
termination of FUSE is rethrown afterwards.

Additionally, exceptions are now logged using python.logging instead of
traceback.print_exc, printing to a possibly closed stderr.

This also fixes OSErrors with negative errnos being forwarded, which
might confuse FUSE.