[fusell] Allow userdata to be passed to constructor
Arbitrary data can be passed as an optional parameter to the FUSELL
constructor. This is then passed to the init and destroy methods.

I've chosen to save the userdata as an instance variable instead of
passing it to fuse_lowlevel_new because that seemed simpler than dealing
with ctypes. We then need to intercept the init and destroy calls to
pass them the correct userdata.